Add routing config for product-details.component
It had to be splited the routing config of token-input.component in
order to specify different child route for product-details.component and
product-list.component. This was neeeded cause the token from these 2
components looks the same and there wouldn't be a way with the token to
decide which component to load.
Also as token-input.component routing config was splited in to 2 now the
way to get the portion from the URL has also changed.