Add method to retrieve product from server
As the retrieval of product takes time, in order to show the product,
we'll use a Subject to show it the time the request has finished. We'll
also save the retrieved object to let us decide whether or not to post
it on a product update, by checkng first if there were actually changes.