Fix validators service
Regex expressions were changed from string to RegExp type as validations
were failing. Also we forced to convert to string data type values from
form controls as a value from a FormControl could be a number, if no
letters were on the FormControl value, and a number varaible doesn't
have either trim() or match() givinng null exception.
On each FormControl validation it was excluded when the field was empty
as this is validated by Validators.required.