Fix bad FormControls validations
Previous way of validating each FormControl had a bug, if one of the
latest fields were fine and a previous one was bad, the PUT request was
still made, cause the flag changed to true at the end.
So we decided to use && operand but this is a short-circuit operand
which when finding the first false wouldn't execute next ones, so we
opted for a non-shor-circuit operand &, but it only works with numbers,
so we had to change it to a number return type the validFormControl().
login.component had the same problem so it was fixed as well.